refactor!: mesh whole slabs rather than pages
This commit is a large refactoring with the aim to remove the variable
sized arrays in Slab and ensure the slab metadata is always less than 1
page long. This is achieved by only storing the bitset in a Slab (which
has a known maximum size base on the max slab size and min slot size)
and moving the shuffle vectors into the meshing pool. The meshing pool
holds one shuffle vector for the current slab being allocated from,
which again has a maximum possible size. Removing the per-page shuffle
vectors reduces the memory overhead significantly and means that
per-page metadata easily fits within a page.

In the furture, memory overhead could be improved by tuning the slab
sizes and slot sizes so that the bitset fill the entire space remaining
of the metadata page and there are enough data pages to saturate the
bitset, though this may only be reasonble to do for larger slot sizes.

+++ src/Span.zig
@@ -18,7 +18,7 @@ const assert = @import("mesh.zig").assert;
 
 const Span = @This();
 
-page_count: u16, // would allow for up to page_size * (2 ^ 16 - 1) total bytes (4KiB under 256MiB for 4KiB pages)
+page_count: usize,
 ptr: [*]align(page_size) u8, // pointer to the mapped memory
 fd: std.os.fd_t, // a valid file descriptor is required for remapping addresses
